A little of refactoring
