+++ cornflow-server/requirements-dev.txt
@@ -22,4 +22,3 @@ sphinxcontrib-jsmath
 sphinxcontrib-qthelp
 sphinxcontrib-serializinghtml
 sphinxcontrib-redoc
-vulture
