+++ libs/core/cornflow_core/cli/tools/schema_generator.py
